move rtp mapper inside controller
